Unification with Kinesis Adv360 layout for the most part.
I think this is now at a point where I will not get massively
confused when switching between the boards.

I'm not yet convinced that this will work for me for coding.
I still think that I could get confused about where the special
characters are. I should look into more example symbol layers
and decide on something that will work both on my favorite
kinesis and on this shitty but portable corne.